Add test for when the --with-stylelint option is passed
How the current test suite is structured is that the generator is part
of the setup and each test method is an assertion. Respecting that
structure, when the generator is going to be invoked with different
options I used a different test class.

Another alternative would be to move the call to run_generator to each
test_* method., but the current structure favors that each test method
tests only one thing.